Refactor FindBandCommand
