+++ src/main/java/seedu/address/logic/parser/FindBandCommandParser.java
@@ -4,7 +4,7 @@ import static seedu.address.logic.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
 
 import java.util.function.Predicate;
 
-import seedu.address.logic.commands.FindBandCommand;
+import seedu.address.logic.commands.band.FindBandCommand;
 import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.band.Band;
 import seedu.address.model.band.BandName;
